[Maintain] Tidy two loose ends in the pool workload layer
Six benchmark subclasses that carry a pool baseline were named after the
workload they extend, differing only by a leading underscore, and each
carried a noqa for N801. Removing the suppression changes nothing -- ruff
does not flag a leading underscore -- so it was never needed. They are now
XxxBenchmarkWorkload, which says what the subclass adds.

No underscore: it would carry nothing here. None of these classes is
imported anywhere, benchmark modules are pytest files nobody imports from,
and across benchmarks/ the split is 61 plain names to 36 underscored, so
plain is both the majority and the one that claims less.

workloads/mean_pooling.py held one class for an op in the pool family, so
the family had two entry files against a convention of one per family.
Folded into workloads/pool.py; the two importers follow.
