+++ benchmarks/ops/bench_mean_pooling.py
@@ -5,8 +5,8 @@ import torch
 
 from benchmarks.benchmark_base import BenchmarkBase, BenchmarkReport
 from tileops.ops import MeanPoolingForwardOp
-from workloads.mean_pooling import MeanPoolingTest
 from workloads.nsa_utils import prepare_chunk_indices
+from workloads.pool import MeanPoolingTest
 
 
 class _MeanPoolingTestBaseline(MeanPoolingTest):
